rename `article.py` to `publication.py`
