+++ scripts/publisher_coverage.py
@@ -11,7 +11,7 @@ from typing import Any, Callable, List, Optional, Union
 
 from fundus import Crawler, PublisherCollection
 from fundus.publishers.base_objects import Publisher, PublisherGroup
-from fundus.scraping.article import Article
+from fundus.scraping.publication import Article
 from fundus.scraping.session import socket_timeout
