+++ src/fundus/__init__.py
@@ -3,7 +3,7 @@ import pathlib
 from langdetect import DetectorFactory
 
 from fundus.publishers import PublisherCollection
-from fundus.scraping.article import Article
+from fundus.scraping.publication import Article
 from fundus.scraping.crawler import CCNewsCrawler, Crawler, CrawlerBase
 from fundus.scraping.filter import Requires
 from fundus.scraping.url import NewsMap, RSSFeed, Sitemap
